#376 kumo fixed start/stop issues; added support for ECS services (#376)
